[S] (~) fix architecture controllers
