--- packages/server/src/infrastructure/web/controllers/auth.controller.ts
@@ -1,56 +1,0 @@
-import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
-import { APIResponse } from "../../../utils/response";
-import { getUserSteamder, login, logout } from "../../../domain/services/authService";
-import { Player } from "../../../models/Players";
-
-/**
- * Handles the steam endpoint.
- *
- * @param request - The FastifyRequest object.
- * @param reply - The FastifyReply object.
- * @returns A Promise that resolves to an API response.
- */
-export const steam = async (request: FastifyRequest, reply: FastifyReply) => {
-  if (!request.user)
-    return APIResponse(reply, null, "logged_in_to_view_profile", 401);
-  return APIResponse(reply, request.user, "logged_in", 200);
-};
-
-/**
- * Handles the Steam callback for authentication.
- *
- * @param fastify - The Fastify instance.
- * @param request - The Fastify request object.
- * @param reply - The Fastify reply object.
- * @returns A promise that resolves to the result of the login operation.
- */
-export const steamCallback = (fastify: FastifyInstance) => async (request: FastifyRequest, reply: FastifyReply) => {
-  return login(fastify, reply, request);
-}
-
-/**
- * Logs out a user.
- *
- * @param fastify - The Fastify instance.
- * @param request - The Fastify request object.
- * @param reply - The Fastify reply object.
- * @returns A Promise that resolves to an API response indicating successful logout.
- */
-export const logoutUser = (fastify: FastifyInstance) => async (request: FastifyRequest, reply: FastifyReply) => {
-  logout(fastify, reply, request);
-  return APIResponse(reply, null, 'logged_out', 200);
-};
-
-/**
- * Retrieves the current user's profile information.
- *
- * @param fastify - The Fastify instance.
- * @returns A Promise that resolves to the user's profile information.
- */
-export const getMe = (fastify: FastifyInstance) => async (request: FastifyRequest, reply: FastifyReply) => {
-  if (!request.user)
-    return APIResponse(reply, null, "logged_in_to_view_profile", 401);
-  const user = request.user as Player & { username: string };
-  const steamder = await getUserSteamder(fastify, BigInt(user.id));
-  return APIResponse(reply, { id: user.id, username: user.username, steamder, avatar_hash: user.avatar_hash }, 'logged_in', 200);
-};
