+++ packages/server/src/infrastructure/web/controllers/debug/index.ts
@@ -1,0 +1,3 @@
+export * from "./getAllEnv"
+export * from "./truncate"
+export * from "./seed"
