+++ packages/server/src/infrastructure/web/controllers/global/index.ts
@@ -1,0 +1,1 @@
+export * from "./stats"
