+++ packages/server/src/infrastructure/web/controllers/index.ts
@@ -1,4 +1,5 @@
-export * from "./auth.controller"
-export * from "./debug.controller"
-export * from "./global.controller"
-export * from "./player.controller"
+export * from "./auth"
+export * from "./debug"
+export * from "./global"
+export * from "./player"
+export * from "./steamder"
