+++ packages/server/src/infrastructure/web/controllers/player/delete.ts
@@ -1,8 +1,8 @@
 import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
-import { APIResponse } from "../../../utils/response";
-import { Player } from "../../../models/Players";
-import { isUserInSteamder } from "../../../domain/services/steamderPlayerService";
-import { deletePlayer, deleteUser } from "../../repositories";
+import { APIResponse } from "../../../../utils/response";
+import { Player } from "../../../../domain/entities";
+import { isUserInSteamder } from "../../../../domain/services/steamderPlayerService";
+import { deletePlayer, deleteUser } from "../../../repositories";
 
 /**
  * Deletes a user account.
