+++ packages/server/src/infrastructure/web/controllers/player/index.ts
@@ -1,0 +1,1 @@
+export * from "./delete"
