+++ packages/server/src/infrastructure/web/routes/auth.route.ts
@@ -1,7 +1,7 @@
 import { FastifyInstance } from 'fastify';
 import fastifyPassport from '@fastify/passport';
 import { isAuthenticated } from '../../../auth/mw';
-import { getMe, logoutUser, steam, steamCallback } from '../controllers/auth.controller';
+import { getMe, logoutUser, steam, steamCallback } from '../controllers/auth';
 import { configureSteamStrategy } from '../../auth/steamStrategy';
 
 export default async function authRouter(fastify: FastifyInstance) {
